requirements.txt
flask
requests
python-dotenv
pygments
openai